--- a/crypto/smartcont/tolk-stdlib/lisp-lists.tolk
+++ b/crypto/smartcont/tolk-stdlib/lisp-lists.tolk
@@ -14,17 +14,18 @@ fun createEmptyList(): tuple
 /// Adds an element to the beginning of lisp-style list.
 /// Note, that it does not mutate the list: instead, it returns a new one (it's a lisp pattern).
 @pure
-fun listPrepend<X>(head: X, tail: tuple): tuple
+fun listPrepend<X>(head: X, tail: tuple?): tuple
     asm "CONS";
 
 /// Extracts the head and the tail of lisp-style list.
 @pure
-fun listSplit<X>(list: tuple): (X, tuple)
+fun listSplit<X>(list: tuple): (X, tuple?)
     asm "UNCONS";
 
 /// Extracts the tail and the head of lisp-style list.
+/// After extracting the last element, tuple is assigned to null.
 @pure
-fun listNext<X>(mutate self: tuple): X
+fun listNext<X>(mutate self: tuple?): X
     asm( -> 1 0) "UNCONS";
 
 /// Returns the head of lisp-style list.
@@ -34,5 +35,5 @@ fun listGetHead<X>(list: tuple): X
 
 /// Returns the tail of lisp-style list.
 @pure
-fun listGetTail(list: tuple): tuple
+fun listGetTail(list: tuple): tuple?
     asm "CDR";