--- a/tolk/tolk-version.h
+++ b/tolk/tolk-version.h
@@ -18,6 +18,6 @@
 
 namespace tolk {
 
-constexpr const char* TOLK_VERSION = "0.8.0";
+constexpr const char* TOLK_VERSION = "0.9.0";
 
 } // namespace tolk