… tolk-tester/tests/a10.tolk
… tolk-tester/tests/allow_post_modification.tolk
… tolk-tester/tests/assignment-tests.tolk
… tolk-tester/tests/c2.tolk
… tolk-tester/tests/cells-slices.tolk
… tolk-tester/tests/codegen_check_demo.tolk
… tolk-tester/tests/dicts-demo.tolk
… tolk-tester/tests/generics-1.tolk
… tolk-tester/tests/imports/use-dicts.tolk
… tolk-tester/tests/indexed-access.tolk
… tolk-tester/tests/inference-tests.tolk
… tolk-tester/tests/invalid-assign-1.tolk
… tolk-tester/tests/invalid-assign-2.tolk
… tolk-tester/tests/invalid-assign-3.tolk
… tolk-tester/tests/invalid-assign-4.tolk
… tolk-tester/tests/invalid-assign-5.tolk
… tolk-tester/tests/invalid-assign-6.tolk
… tolk-tester/tests/invalid-assign-7.tolk
… tolk-tester/tests/invalid-call-10.tolk
… tolk-tester/tests/invalid-generics-1.tolk
… tolk-tester/tests/invalid-generics-13.tolk
… tolk-tester/tests/invalid-generics-14.tolk
… tolk-tester/tests/invalid-generics-7.tolk
… tolk-tester/tests/invalid-mutate-18.tolk
… tolk-tester/tests/invalid-mutate-19.tolk
… tolk-tester/tests/invalid-mutate-20.tolk
… tolk-tester/tests/invalid-never-1.tolk
… tolk-tester/tests/invalid-typing-11.tolk
… tolk-tester/tests/invalid-typing-14.tolk
… tolk-tester/tests/invalid-typing-15.tolk